nba_fantasy_waiver_advisor/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
nba_api==1.7.0
numpy==1.26.4
requests==2.32.3
urllib3==2.3.0
python-dotenv==1.0.0
espn-api==0.8.2